Variable not large enough to hold large initial offsets.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@1193 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/input_node_data_writer.h
@@ -105,7 +105,7 @@ private:
                             int invalid_samples_per_block, int block_size);
   void write_end_of_stream(Data_writer_sptr writer);
 
-  int64_t write_initial_invalid_data(Writer_struct &data_writer, int byte_offset);
+  int64_t write_initial_invalid_data(Writer_struct &data_writer, int64_t byte_offset);
   uint64_t total_data_written_;
   int block_size;
